+++ src/zodbpickle/__init__.py
@@ -1,8 +1,11 @@
-# Make a package.
+import sys
 
-class binary(bytes):
-    """Mark a given string as explicitly binary
+if sys.version_info[0] < 3:
+    class binary(bytes):
+        """Mark a given string as explicitly binary
 
-    I.e., it should be unpickled as bytes on Py3k, rather than being
-          forcibly promoted to unicode.
-    """
+        I.e., it should be unpickled as bytes on Py3k, rather than being
+            forcibly promoted to unicode.
+        """
+else:
+    binary = bytes