+++ CHANGES.rst
@@ -1,5 +1,5 @@
-CHANGES
-=======
+``zodbpickle`` Changelog
+========================
 
 0.4 (unreleased)
 ----------------